Move base64 functionality to 'hpke' module
