+++ lib/bytes/include/bytes/bytes.h
@@ -115,16 +115,4 @@ to_hex(const bytes& data);
 bytes
 from_hex(const std::string& hex);
 
-std::string
-to_base64(const bytes& data);
-
-std::string
-to_base64url(const bytes& data);
-
-bytes
-from_base64(const std::string& enc);
-
-bytes
-from_base64url(const std::string& enc);
-
 } // namespace bytes_ns
